try2

+++ Scatt3D/scatteringProblem.py
@@ -477,9 +477,9 @@ class Scatt3DProblem():
         :param ref: If 1, deletes ref fields. If 2, deletes dut fields. If 0, deletes both.
         '''
         if(self.comm.rank == self.model_rank): ## presumably it's fine to just delete with 1 process
-            if(ref==0 or ref==1):
+            if(ref==0 or ref==1 and hasattr(self, self.dataFolder+self.name+'Ref.Solutions')):
                 shutil.rmtree(self.dataFolder+self.name+'Ref.Solutions')
-            if(ref==0 or ref==2):
+            if(ref==0 or ref==2 and hasattr(self, self.dataFolder+self.name+'Dut.Solutions')):
                 shutil.rmtree(self.dataFolder+self.name+'Dut.Solutions')
